use macro TYPE_DISPATCH_ALL
